Accept EDI and LTER data package inputs
The functions download_d1_data_pkg() and read_d1_files() fails to download and read EDI and LTER data package contents because of a misidentified pid identifier for data package "reports". These reports (.xml) are listed as "DATA" and are sent to table reading functions that throw an error. This enhancement removes the report from the list of pids and allows download and reading to proceed without error.